change:directory configuration for users and administrators
