+++ resources/ts/features/store.ts
@@ -9,9 +9,9 @@ import {
   REGISTER,
 } from "redux-persist";
 import storage from "redux-persist/lib/storage/session";
-import authReducer from "./auth/authSlice";
-import profileReducer from "./profile/profileSlice";
-import deskReducer from "./desk/deskSlice";
+import authReducer from "./user/auth/authSlice";
+import profileReducer from "./user/profile/profileSlice";
+import deskReducer from "./user/desk/deskSlice";
 import commonReducer from "./common/commonSlice";
 
 // redux-presistの設定
